backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.3
httpx==0.26.0
requests==2.31.0
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.8.0
boto3==1.34.34
reportlab==4.0.9

